Add play/pause buttons, don't smooth all the way to the beginning
Play pause is just getting some more ways to do stuff that was already in the game. Play was previously only via spacebar and pause could be done by dragging timescale to 0. Now there is explicit UI functionality for both.

Also made the lines not smoothen all the way back, which means you can draw lines as long as you want and the game shouldn't lag too much. Instead only the last 15 points are used for smoothing which is plenty...